+++ firmware/test/test_mainUnity.cpp
@@ -19,6 +19,10 @@ void test_scoped_analog_provider_nesting();
 void test_sequence_provider_cycles_values();
 void test_set_provider_returns_previous();
 void test_digital_provider_overrides_matrix_reads();
+void test_parse_basic_sysex_template();
+void test_parse_14bit_sysex_template();
+void test_parse_rejects_bad_template();
+void test_config_mutation_during_stream_stays_valid();
 void test_bulk_config_assembler_handles_chunks();
 void test_bulk_config_assembler_detects_overflow();
 void test_format_ack_includes_checksum_and_seq();
@@ -34,9 +38,16 @@ void test_pitch_bend();
 void test_send_nrpn();
 void test_receive_nrpn();
 void test_send_sysex();
+void test_pot_burst_keeps_cc_counters_honest();
+void test_long_sysex_payload_round_trips();
 void test_drop_unsupported_usb_type();
 void test_usb_clock_tick_advances_counter();
 void test_generate_clock_tick_advances_counter();
+void test_handle_sysex_drops_oversize();
+void test_config_manager_wipes_legacy_slot_stride();
+void test_pot_burst_keeps_cc_counters_honest();
+void test_long_sysex_payload_round_trips();
+void test_config_mutation_during_stream_stays_valid();
 #endif
 
 void setup() {
@@ -45,6 +56,10 @@ void setup() {
     RUN_TEST(test_sequence_provider_cycles_values);
     RUN_TEST(test_set_provider_returns_previous);
     RUN_TEST(test_digital_provider_overrides_matrix_reads);
+    RUN_TEST(test_parse_basic_sysex_template);
+    RUN_TEST(test_parse_14bit_sysex_template);
+    RUN_TEST(test_parse_rejects_bad_template);
+    RUN_TEST(test_config_mutation_during_stream_stays_valid);
     RUN_TEST(test_bulk_config_assembler_handles_chunks);
     RUN_TEST(test_bulk_config_assembler_detects_overflow);
     RUN_TEST(test_format_ack_includes_checksum_and_seq);
@@ -65,13 +80,20 @@ void setup() {
     RUN_TEST(test_program_change);
     RUN_TEST(test_aftertouch);
     RUN_TEST(test_mod_wheel);
+    RUN_TEST(test_pot_burst_keeps_cc_counters_honest);
     RUN_TEST(test_pitch_bend);
     RUN_TEST(test_send_nrpn);
     RUN_TEST(test_receive_nrpn);
     RUN_TEST(test_send_sysex);
+    RUN_TEST(test_long_sysex_payload_round_trips);
     RUN_TEST(test_drop_unsupported_usb_type);
+    RUN_TEST(test_handle_sysex_drops_oversize);
     RUN_TEST(test_usb_clock_tick_advances_counter);
     RUN_TEST(test_generate_clock_tick_advances_counter);
+    RUN_TEST(test_config_manager_wipes_legacy_slot_stride);
+    RUN_TEST(test_pot_burst_keeps_cc_counters_honest);
+    RUN_TEST(test_long_sysex_payload_round_trips);
+    RUN_TEST(test_config_mutation_during_stream_stays_valid);
 #endif
     UNITY_END();
 }